content_analysis_agent/requirements.txt
# Content Analysis Agent Dependencies

# Core dependencies
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.28.0
tenacity>=8.2.0

# Content processing
trafilatura>=1.6.0
markdownify>=0.11.0
textstat>=0.7.0
nltk>=3.8.0
python-dateutil>=2.8.0

# Configuration and data handling
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Data analysis and statistics
pandas>=2.0.0
numpy>=1.24.0

# Optional: For enhanced web scraping if needed
selenium>=4.15.0
webdriver-manager>=4.0.0